# docs/swagger.yaml
basePath: /
definitions:
...
  model.CreateProjectRequest:
    properties:
      projectName:
        type: string
      userID:
        type: string
    type: object
  model.CreateProjectResponse:
    properties:
      code:
        format: int32
        type: integer
      error:
        type: boolean
      message:
        type: string
    type: object
...
  model.DeleteProjectResponse:
    properties:
      code:
        format: int32
        type: integer
      error:
        type: boolean
      message:
        type: string
    type: object
...
  model.GetProjectDetailResponse:
    properties:
      code:
        format: int32
        type: integer
      data:
        $ref: '#/definitions/model.Project'
      error:
        type: boolean
      message:
        type: string
    type: object
...
  model.ListProjectResponse:
    properties:
      code:
        format: int32
        type: integer
      data:
        items:
          $ref: '#/definitions/model.Project'
        type: array
      error:
        type: boolean
      message:
        type: string
      pagination:
        $ref: '#/definitions/model.PaginationResponse'
    type: object
...
  model.PaginationResponse:
    properties:
      limit:
        format: int32
        type: integer
      page:
        format: int32
        type: integer
      totalPages:
        format: int32
        type: integer
      totalRows:
        format: int64
        type: integer
    type: object
  model.Project:
    properties:
      createdAt:
        type: string
      createdById:
        type: integer
      projectID:
        type: integer
      projectName:
        type: string
      status:
        type: integer
      statusDesc:
        type: string
      updatedAt:
        type: string
      updatedById:
        type: integer
    type: object
...
  model.UpdateProjectRequest:
    properties:
      options:
        type: string
      projectID:
        type: string
      projectName:
        type: string
      status:
        format: int32
        type: integer
      statusDesc:
        type: string
      userID:
        type: string
    type: object
  model.UpdateProjectResponse:
    properties:
      code:
        format: int32
        type: integer
      error:
        type: boolean
      message:
        type: string
    type: object
...
  utils.APIErrorResponse:
    properties:
      Code:
        type: integer
      Error:
        type: boolean
      Message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the RAWUH Service API.
  title: RAWUH Service API
  version: "1.0"
paths:
...
  /project:
    post:
      consumes:
      - application/json
      description: Create a new project
      parameters:
      - description: CreateProjectRequest
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreateProjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      security:
      - Bearer: []
      summary: Create a project
      tags:
      - project
  /project/{project_id}:
    delete:
      consumes:
      - application/json
      description: Delete project by ID
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeleteProjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      summary: Delete a project
      tags:
      - project
    get:
      consumes:
      - application/json
      description: Retrieve project details by ID
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetProjectDetailResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      summary: Get project detail
      tags:
      - project
    put:
      consumes:
      - application/json
      description: Update project by ID
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: UpdateProjectRequest
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UpdateProjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      summary: Update a project
      tags:
      - project
  /project/list:
    get:
      consumes:
      - application/json
      description: Get paginated list of projects
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListProjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      security:
      - Bearer: []
      summary: List projects
      tags:
      - project
...
swagger: "2.0"